Add seal verification method to block header
