+++ subspace/crates/subspace-farmer/src/bin/subspace-farmer/commands/farm.rs
@@ -1,6 +1,7 @@
 use crate::commands::shared::network::{NetworkArgs, configure_network};
 use crate::commands::shared::{DiskFarm, PlottingThreadPriority, derive_libp2p_keypair};
 use crate::utils::shutdown_signal;
+use ab_core_primitives::ed25519::Ed25519PublicKey;
 use ab_erasure_coding::ErasureCoding;
 use anyhow::anyhow;
 use async_lock::{Mutex as AsyncMutex, RwLock as AsyncRwLock, Semaphore};
@@ -51,7 +52,6 @@ use subspace_farmer_components::reading::ReadSectorRecordChunksMode;
 use subspace_metrics::{RegistryAdapter, start_prometheus_metrics_server};
 use subspace_networking::utils::piece_provider::PieceProvider;
 use subspace_proof_of_space::Table;
-use subspace_verification::ed25519::Ed25519PublicKey;
 use tracing::{Instrument, error, info, info_span, warn};
 
 /// Get piece retry attempts number.
