+++ subspace/crates/subspace-farmer/src/plotter/cpu.rs
@@ -6,6 +6,7 @@ use crate::plotter::cpu::metrics::CpuPlotterMetrics;
 use crate::plotter::{Plotter, SectorPlottingProgress};
 use crate::thread_pool_manager::PlottingThreadPoolManager;
 use crate::utils::AsyncJoinOnDrop;
+use ab_core_primitives::ed25519::Ed25519PublicKey;
 use ab_core_primitives::sectors::SectorIndex;
 use ab_erasure_coding::ErasureCoding;
 use async_lock::{Mutex as AsyncMutex, Semaphore, SemaphoreGuardArc};
@@ -34,7 +35,6 @@ use subspace_farmer_components::plotting::{
     encode_sector, write_sector,
 };
 use subspace_proof_of_space::Table;
-use subspace_verification::ed25519::Ed25519PublicKey;
 use tokio::task::yield_now;
 use tracing::{Instrument, warn};
