+++ subspace/crates/subspace-farmer/src/plotter/pool.rs
@@ -1,6 +1,7 @@
 //! Pool plotter
 
 use crate::plotter::{Plotter, SectorPlottingProgress};
+use ab_core_primitives::ed25519::Ed25519PublicKey;
 use ab_core_primitives::sectors::SectorIndex;
 use async_trait::async_trait;
 use event_listener::Event;
@@ -10,7 +11,6 @@ use std::any::type_name_of_val;
 use std::pin::pin;
 use std::time::Duration;
 use subspace_farmer_components::FarmerProtocolInfo;
-use subspace_verification::ed25519::Ed25519PublicKey;
 use tracing::{error, trace};
 
 /// Pool plotter.
