+++ subspace/crates/subspace-verification/Cargo.toml
@@ -15,27 +15,16 @@ include = [
 
 [dependencies]
 ab-core-primitives = { workspace = true }
-derive_more = { workspace = true, features = ["deref", "from", "into"] }
 ed25519-zebra = { workspace = true }
-hex = { workspace = true, optional = true }
 parity-scale-codec = { workspace = true, features = ["bytes", "derive", "max-encoded-len"], optional = true }
 scale-info = { workspace = true, features = ["derive"], optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
-serde-big-array = { workspace = true }
 
 [features]
-alloc = [
-    "hex?/alloc",
-    "serde/alloc",
-]
 scale-codec = [
     "dep:parity-scale-codec",
     "dep:scale-info",
-    "alloc",
 ]
 serde = [
-    "alloc",
-    "dep:hex",
     "dep:serde",
-    "hex/serde",
 ]
